# api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: Assisted Migration Agent API
  version: v2
servers:
  - url: /api/v2
...
paths:
...
  /agent:
    get:
      tags: [Agent]
      summary: Get agent status
      operationId: getAgentStatus
      responses:
        '200':
          description: Agent status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatus'
        '500':
          description: Internal server error
    post:
      tags: [Agent]
      summary: Change agent mode
      operationId: setAgentMode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentModeRequest'
      responses:
        '200':
          description: Mode changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatus'
        '400':
          description: Invalid request
        '409':
          description: Conflict
        '500':
          description: Internal server error

  # ── Collections ────────────────────────────────────────────────────────
...
  /collectors:
    post:
      tags: [Collectors]
      summary: Start a collection
      operationId: startCollector
      responses:
        '202':
          description: Collection started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectorStatus'
        '400':
          description: Invalid request
        '409':
          description: Collection already in progress
        '500':
          description: Internal server error
    get:
      tags: [Collectors]
      summary: List all collectors with status
      operationId: listCollectors
      responses:
        '200':
          description: List of collectors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectorListResponse'
        '500':
          description: Internal server error
  /collectors/{id}:
    get:
      tags: [Collectors]
      summary: Get collector status
      operationId: getCollectorStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Collector ID
          schema:
            type: string
      responses:
        '200':
          description: Collector status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectorStatus'
        '404':
          description: Collector not found
        '500':
          description: Internal server error
    delete:
      tags: [Collectors]
      summary: Stop collector
      operationId: stopCollector
      parameters:
        - name: id
          in: path
          required: true
          description: Collector ID
          schema:
            type: string
      responses:
        '204':
          description: Collector stopped
        '404':
          description: Collector not found
        '500':
          description: Internal server error

  # ── Inventories ─────────────────────────────────────────────────────────
...
  /collections/{id}/applications:
    get:
      tags: [Applications]
      summary: List detected applications in a collection
      operationId: listApplications
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: List of detected applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationListResponse'
        '404':
          description: Collection not found
        '500':
          description: Internal server error

  # ── Groups ─────────────────────────────────────────────────────────────
...
  /credentials:
    put:
      tags: [Credentials]
      summary: Store vCenter credentials
      operationId: putCredentials
      description: >-
        Validates credentials against vSphere, encrypts with disk-based key,
        and stores in the local database as the active credential set.
        A subsequent PUT overwrites the previously stored credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VcenterCredentials'
      responses:
        '201':
          description: Credentials created and stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialStatus'
        '200':
          description: Credentials validated and stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialStatus'
        '400':
          description: Invalid credentials or unreachable vCenter
        '500':
          description: Internal server error
    get:
      tags: [Credentials]
      summary: Get stored credential status
      operationId: getCredentials
      description: >-
        Returns stored credential metadata (URL only). Never exposes
        username or password. This endpoint does not revalidate credentials
        against vSphere.
      responses:
        '200':
          description: Credential status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialStatus'
        '404':
          description: No credentials stored
        '500':
          description: Internal server error
    delete:
      tags: [Credentials]
      summary: Delete stored credentials
      operationId: deleteCredentials
      description: Removes all stored vCenter credentials from the database.
      responses:
        '204':
          description: Credentials deleted
        '500':
          description: Internal server error
  /credentials/capabilities:
    get:
      tags: [Credentials]
      summary: Check vSphere operation capabilities
      operationId: getCredentialCapabilities
      description: >-
        Checks vSphere privileges for all operations (collector, inspector,
        forecaster) using stored credentials. Single live vSphere check, no caching.
      responses:
        '200':
          description: Capability status per operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityStatus'
        '404':
          description: No credentials stored
        '500':
          description: Internal server error

  # ── Version ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CollectorStatus:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          description: Collector identifier
        status:
          type: string
          enum:
            - connecting
            - collecting
            - parsing
            - collected
            - error
          x-enum-varnames:
            - CollectorStatusStatusConnecting
            - CollectorStatusStatusCollecting
            - CollectorStatusStatusParsing
            - CollectorStatusStatusCollected
            - CollectorStatusStatusError
        error:
          type: string
          description: Error message when status is error
    CollectorListResponse:
      type: object
      required:
        - collectors
      properties:
        collectors:
          type: array
          items:
            $ref: '#/components/schemas/CollectorStatus'

    # ── Inventories ──────────────────────────────────────────────────────
...
    VcenterCredentials:
      required:
        - url
        - username
        - password
      properties:
        url:
          type: string
          format: uri
          description: vCenter URL
          x-oapi-codegen-extra-tags:
            binding: "required,url"
        username:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
        password:
          type: string
          format: password
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
        cacert:
          type: string
          description: >-
            PEM-encoded CA certificate bundle for verifying the vCenter TLS
            certificate. Mutually exclusive with skipTls.
          x-oapi-codegen-extra-tags:
            binding: "omitempty"
        skipTls:
          type: boolean
          description: >-
            When true, TLS certificate verification is skipped. Must not be set
            to true alongside cacert — that combination returns 400. When
            omitted: if cacert is absent the connection skips TLS verification
            (backwards compatibility); if cacert is provided the connection
            verifies TLS using that certificate — skipTls need not be sent.
    CredentialStatus:
      type: object
      required:
        - url
        - username
        - valid
      properties:
        url:
          type: string
          description: vCenter URL the credentials belong to
        username:
          type: string
          description: vCenter username
        valid:
          type: boolean
          description: Whether credentials are stored for the returned URL.
    OperationCapability:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: Whether stored credentials have sufficient privileges
        missingPrivileges:
          type: array
          items:
            type: string
          description: Specific vSphere privileges that are missing
    CapabilityStatus:
      type: object
      required:
        - capabilities
      properties:
        capabilities:
          type: object
          required:
            - collector
            - inspector
            - forecaster
          properties:
            collector:
              $ref: '#/components/schemas/OperationCapability'
            inspector:
              $ref: '#/components/schemas/OperationCapability'
            forecaster:
              $ref: '#/components/schemas/OperationCapability'

    # ── Virtual Machines ─────────────────────────────────────────────────
...
    ApplicationListResponse:
      type: object
      required:
        - applications
      properties:
        applications:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationOverview'
    ApplicationOverview:
      type: object
      required:
        - name
        - description
        - vmCount
        - vms
      properties:
        name:
          type: string
          description: Application name
        description:
          type: string
          description: Application description
        vmCount:
          type: integer
          description: Number of VMs running this application
        vms:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationVM'
    ApplicationVM:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: VM ID
        name:
          type: string
          description: VM name

    # ── VM Operations ───────────────────────────────────────────────────
...
    AgentStatus:
      type: object
      required:
        - mode
        - consoleConnection
      properties:
        mode:
          type: string
          enum:
            - connected
            - disconnected
          x-enum-varnames:
            - AgentStatusModeConnected
            - AgentStatusModeDisconnected
          description: Target mode for the agent
        consoleConnection:
          type: object
          required:
            - status
          properties:
            status:
              type: string
              enum:
                - connected
                - disconnected
              x-enum-varnames:
                - ConsoleConnectionStatusConnected
                - ConsoleConnectionStatusDisconnected
              description: Current console connection status
            error:
              type: string
              description: Connection error description
    AgentModeRequest:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum:
            - connected
            - disconnected
          x-oapi-codegen-extra-tags:
            binding: "required,oneof=connected disconnected"

    # ── Version ──────────────────────────────────────────────────────────
...